Finished new type of cell tracker
Tracks cells thru timepoints. Still needs a few tweaks:

Aligned to reference cell but not reference axis, so far away cells move a lot when the germarium rotates, which causes miscalculations in the cell tracker.

Cells are not yet split up (e.g. when the cell has >5 slices)

Tiny bit of lag. Does it matter???